--- a/src/main/resources/db/migration/V3__Add_clinic_activity_logs_numeric_value_index.sql
+++ b/src/main/resources/db/migration/V3__Add_clinic_activity_logs_numeric_value_index.sql
@@ -1,2 +1,2 @@
 -- Add index on numeric_value column to improve query performance
-CREATE INDEX IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs (numeric_value);
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_clinic_activity_logs_numeric_value ON clinic_activity_logs (numeric_value);