This PR addresses a critical performance issue in the clinic activity logs query system where queries filtering on the numeric_value column were experiencing a 7,900x performance degradation (4.69s vs 577.3μs).

Changes made:
1. Added a B-tree index on the numeric_value column of the clinic_activity_logs table
2. Created a migration script to add the index safely using CONCURRENTLY option

The index will improve query performance for the following query pattern:
```sql
SELECT id, activity_type, numeric_value, event_timestamp, status_flag, payload 
FROM clinic_activity_logs 
WHERE numeric_value = ?
```

Affected endpoint: /api/clinic-activity/query-logs
Primary issue ID: 5aa1fb7a-3656-11f0-8db3-ca59c7e8e81d

Expected performance improvement:
- Before: 4.69s query execution time
- After: ~577.3μs query execution time (7,900x improvement)

Testing:
- The migration has been tested in a development environment
- The CONCURRENTLY option ensures no table locks during index creation
- Query performance has been verified with EXPLAIN ANALYZE